coordinator/pyproject.toml
[project]
name = "wardenhub-coordinator"
version = "0.1.0"
description = "WardHub security auditing coordinator"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pyyaml>=6.0",
    "prometheus-client>=0.20",
    "httpx>=0.27",
    "apscheduler>=3.10,<4.0",
    "structlog>=24.0",
    "pydantic>=2.0",
]

[project.scripts]
wardenhub-coordinator = "wardenhub_coordinator.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -m 'not integration'"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]
